feat(env/sysctl): add support for user.max_user_namespaces
- Implemented sysctl function `MaxUserNamespaces()` to retrieve and parse `/proc/sys/user/max_user_namespaces`.
- Added corresponding structures and output fields in `env/sysctl/print.go` and `env/sysctl/sysctl.go`.
- Moved previous logic from `prerequisite/proc` to `prerequisite/sysctl/user`.
- Updated module dependency `github.com/ctrsploit/sploit-spec` to `v0.6.1-rc1`.
- Adjusted related import paths accordingly and cleaned up obsolete files.
