+++ vul/cve-2022-0492/vul.go
@@ -4,8 +4,8 @@ import (
 	"github.com/ctrsploit/ctrsploit/prerequisite/apparmor"
 	"github.com/ctrsploit/ctrsploit/prerequisite/cgroups"
 	"github.com/ctrsploit/ctrsploit/prerequisite/kernel/cve-2022-0492"
-	procUser "github.com/ctrsploit/ctrsploit/prerequisite/proc/sys/user"
 	"github.com/ctrsploit/ctrsploit/prerequisite/syscall"
+	procUser "github.com/ctrsploit/ctrsploit/prerequisite/sysctl/user"
 	"github.com/ctrsploit/ctrsploit/prerequisite/user"
 	"github.com/ctrsploit/sploit-spec/pkg/app"
 	"github.com/ctrsploit/sploit-spec/pkg/exeenv"
